% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diabetes.R
\docType{data}
\name{Diabetes}
\alias{Diabetes}
\title{Sub-data of the 7th Korea National Health and Nutrition Examination Survey in 2016.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4367 rows and 14 columns.}
\source{
\href{https://knhanes.cdc.go.kr/knhanes/sub03/sub03_02_02.do}{KCDC}
}
\usage{
data("Diabetes")
}
\description{
Data from Korea Centers for Disease Control and Prevention.
Only have diabetes related variables of 4,367 people.
}
\examples{
data("Diabetes")
}
\references{
The Third Korea National Health and Nutrition Examination Survey,
2016, Korea Centers for Disease Control and Prevention
}
\keyword{datasets}
