% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diabetes_Models.R
\docType{data}
\name{Diabetes_Models}
\alias{Diabetes_Models}
\title{Pre-trained models for diabetes prediction.}
\format{An object of class \code{caretList} of length 6.}
\usage{
data("Diabetes_Models")
}
\description{
Using 7th Korea National Health and Nutrition Examination Survey in 2016,
build pre-trained model for diabetes predicting. 6 models are used and
average AUC is about 0.72. But this pre-trained models have terrible performance
for predicting not normal(F1 0.4), so do not take the result seriously.
}
\examples{
data("Diabetes_Models")
}
\keyword{datasets}
