% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes_predict.R
\name{diabetes_predict}
\alias{diabetes_predict}
\title{Predicting risk probability of get Diabetes.}
\usage{
diabetes_predict(mode, dataframe = NULL)
}
\arguments{
\item{mode}{select mode, "insert" or "dataframe".}

\item{dataframe}{if argument mode is "dataframe", dataframe argument should not be NULL.}
}
\value{
probabilities of get disease.
}
\description{
mle_gamma function will calculate mle estimate for the shape parameter
"k" and the scale parameter "theta" by using given sample. Since there is no closed form of the mle for gamma distribution, mle_gamma used uniroot function from stats package.
}
\details{
If argument "mode" is dataframe, you should be set column name correctly.\cr
 Explanation of columns will followed.\cr
 "sex_m": gender of sample. if sample is male, set 1 else, 0.\cr
 "age": age of sample.\cr
 "A_WV": minutes of vigorous physical activities at work per week.
 "A_WM": minutes of moderate physical activities at work per week.
 "A_HV": minutes of vigorous physical activities as hobby per week.
 "A_HM": minutes of moderate physical activities as hobby per week.
 "A_MV": minutes of moving place to place with walking or biking and so on per week.
 "A_ST": minutes of sitting or lying down per day(usually).
 "A_WD": minutes of walking per week.
 "S_CT": amount of smoking per day.\cr
 "D_YF": frequency of drinking per month.\cr
 "D_AT": amount of drinking per once.\cr
 "D_DD": frequency of dringking badly(over 7 glasses in soju).
}
\examples{
  data <- data.frame(
    sex_m = c(0, 1),
    age = c(20, 29),
    A_WV = c(0, 360),
    A_WM = c(0, 1500),
    A_HV = c(0, 0),
    A_HM = c(240, 0),
    A_MV = c(50, 80),
    A_ST = c(960, 240),
    A_WD = c(40, 80),
    S_CT = c(0, 10),
    D_YF = c(1, 3),
    D_AT = c(3, 7),
    D_DD = c(0, 1)
  )
  test <- diabetes_predict("dataframe", data)
}
