% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diabetes_Stack.R
\docType{data}
\name{Diabetes_Stack}
\alias{Diabetes_Stack}
\title{Pre-trained ensemble model for diabetes prediction.}
\format{An object of class \code{caretStack} of length 3.}
\usage{
data("Diabetes_Stack")
}
\description{
Using 7th Korea National Health and Nutrition Examination Survey in 2016,
build pre-trained ensemble model for diabetes predicting. 6 models are used and
average AUC is about 0.72. But this pre-trained models have terrible performance
for predicting not normal(F1 0.4), so do not take the result seriously.
}
\examples{
data("Diabetes_Stack")
}
\keyword{datasets}
