% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypertension_Models.R
\docType{data}
\name{Hypertension_Models}
\alias{Hypertension_Models}
\title{Pre-trained models for hypertension prediction.}
\format{An object of class \code{caretList} of length 6.}
\usage{
data("Hypertension_Models")
}
\description{
Using 7th Korea National Health and Nutrition Examination Survey in 2016,
build pre-trained model for hypertension predicting. 6 models are used and
average AUC is about 0.79
}
\examples{
data("Hypertension_Models")
}
\keyword{datasets}
