% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertension_predict.R
\name{hypertension_predict}
\alias{hypertension_predict}
\title{Predicting risk probability of get Hypertension.}
\usage{
hypertension_predict(mode, dataframe = NULL)
}
\arguments{
\item{mode}{select mode, "insert" or "dataframe".}

\item{dataframe}{if argument mode is "dataframe", dataframe argument should not be NULL.}
}
\value{
probabilities of get disease.
}
\description{
mle_gamma function will calculate mle estimate for the shape parameter
"k" and the scale parameter "theta" by using given sample. Since there is no closed form of the mle for gamma distribution, mle_gamma used uniroot function from stats package.
}
\details{
If argument "mode" is dataframe, you should be set column name correctly.\cr
 Explanation of columns will followed.\cr
 "sex_m": gender of sample. if sample is male, set 1 else, 0.\cr
 "age": age of sample.\cr
 "S_CT": amount of smoking per day.\cr
 "D_YF": frequency of drinking per month.\cr
 "D_AT": amount of drinking per once.\cr
 "D_DD": frequency of dringking badly(over 7 glasses in soju).\cr
 "E_BR": frequency of having breakfast per week.\cr
 "E_LN": frequency of having lunch per week.\cr
 "E_DN": frequency of having dinner per week.
}
\examples{
  data <- data.frame(
    sex_m = c(0, 1),
    age = c(20, 29),
    S_CT = c(0, 10),
    D_YF = c(1, 3),
    D_AT = c(3, 7),
    D_DD = c(0, 1),
    E_BR = c(3, 0),
    E_LN = c(7, 5),
    E_DN = c(3, 0)
  )
  test <- hypertension_predict("dataframe", data)
}
