% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypertension_Stack.R
\docType{data}
\name{Hypertension_Stack}
\alias{Hypertension_Stack}
\title{Pre-trained ensemble model for hypertension prediction.}
\format{An object of class \code{caretStack} of length 3.}
\usage{
data("Hypertension_Stack")
}
\description{
Using 7th Korea National Health and Nutrition Examination Survey in 2016,
build pre-trained ensemble stacking model for hypertension predicting.
6 models which in Hypertension_Models are used and average AUC is about 0.79
}
\examples{
data("Hypertension_Stack")
}
\keyword{datasets}
