% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypertensions.R
\docType{data}
\name{Hypertensions}
\alias{Hypertensions}
\title{Sub-data of the 7th Korea National Health and Nutrition Examination Survey in 2016.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5096 rows and 10 columns.}
\source{
\href{https://knhanes.cdc.go.kr/knhanes/sub03/sub03_02_02.do}{KCDC}
}
\usage{
data("Hypertensions")
}
\description{
Data from Korea Centers for Disease Control and Prevention.
Only have hypertension related variables of 4,367 people.
}
\examples{
data("Hypertensions")
}
\references{
The Third Korea National Health and Nutrition Examination Survey,
2016, Korea Centers for Disease Control and Prevention
}
\keyword{datasets}
